refactor: Refactoring packages for mejaservice enums
