+++ src/test/java/id/ac/ui/cs/advprog/ohioorder/meja/repository/MejaRepositoryTest.java
@@ -1,6 +1,6 @@
 package id.ac.ui.cs.advprog.ohioorder.meja.repository;
 
-import id.ac.ui.cs.advprog.ohioorder.enums.MejaStatus;
+import id.ac.ui.cs.advprog.ohioorder.meja.enums.MejaStatus;
 import id.ac.ui.cs.advprog.ohioorder.meja.model.Meja;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
